Bug 1046822 — Replace the hawkMiddleware with express-hawk.
