+++ msisdn-gateway/errno.json
@@ -3,7 +3,6 @@
  "BADJSON": 106,
  "INVALID_PARAMETERS": 107,
  "MISSING_PARAMETERS": 108,
- "INVALID_REQUEST_SIG": 109,
  "INVALID_AUTH_TOKEN": 110,
  "EXPIRED": 111,
  "LENGTH_MISSING": 112,
