+++ package.json
@@ -19,6 +19,7 @@
     "cors": "2.2.0",
     "express": "3.x",
     "express-logging": "git://github.com/mozilla-services/express-logging.git",
+    "express-hawkauth": "0.2.0",
     "hawk": "2.2.3",
     "hiredis": "0.1.17",
     "hkdf": "0.0.2",
